+++ src/spycipdb/clis/cli.py
@@ -23,6 +23,7 @@ from spycipdb.clis import(
     cli_jc,
     cli_cs,
     cli_saxs,
+    cli_rh,
     cli_smfret,
     )
 
@@ -38,6 +39,7 @@ Core back-calculator functions:
     * {cli_jc._name}
     * {cli_cs._name}
     * {cli_saxs._name}
+    * {cli_rh._name}
     * {cli_smfret._name}
 """
 
@@ -60,6 +62,7 @@ libcli.add_subparser(subparsers, cli_noe)
 libcli.add_subparser(subparsers, cli_jc)
 libcli.add_subparser(subparsers, cli_cs)
 libcli.add_subparser(subparsers, cli_saxs)
+libcli.add_subparser(subparsers, cli_rh)
 libcli.add_subparser(subparsers, cli_smfret)
 
 def load_args():